DX-1071: Add --env=env argument
Since the attribute Jekyll.env is read-only, the only way to affect it is
either through brittle and hacky monkey-patching (which I couldn't get
to work consistently) or by setting the JEKYLL_ENV environment variable,
the argument had to be implemented in entrypoint.sh.

I also changed `bundle config set gemfile` to a process-scoped variable,
so it's a bit less tiresome to test entrypoint.sh locally.